+++ src/runtime/window/window_system.hpp
@@ -7,8 +7,8 @@ NAMESPACE_BEGIN(window)
 
 struct Config
 {
-    u16 width = Game::DEFAULT_WIDTH;
-    u16 height = Game::DEFAULT_HEIGHT;
+    u32 width = Game::DEFAULT_WIDTH;
+    u32 height = Game::DEFAULT_HEIGHT;
     const char* title = Game::WINDOW_TITLE;
     bool use_vulkan = true;
 };
